crud new

+++ CRUD/index.php
@@ -4,7 +4,17 @@ $info = '';
 $task = $_GET['task'] ?? 'report';
 $error = $_GET['error'] ?? '0';
 
+if('edit' == $task) {
+    if(!hasPrivilege()) {
+        header('location:index.php?task=report');
+    }
+}
+
 if( 'delete' == $task) {
+    if(!isAdmin()) {
+        header('location:index.php?task=report');
+        return;
+    }
     $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
     if( $id>0) {
         deleteStudent($id);
@@ -12,6 +22,10 @@ if( 'delete' == $task) {
     }
 }
 if('seed' == $task) {
+    if(!isAdmin()) {
+        header('location:index.php?task=report');
+        return;
+    }
     seed();
     $info = "Seeding is complete";
 }
